Make version info an option; add main repo mode
Bah, version info doesn't matter with Java. I might change it up later,
but for now, it's only GCC and it's an option.

`-o` triggers 'aqa-computing' mode, which means your work is in main
repo `aqa-computing` rather than its own repo. The latter has become
default behaviour for me, so it's been depromoted to an option. What it
does is simply rewrite the GitHub URL to work for subdirectories (to a
very limited extent).